+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- #320: Method for `Pattern`: `compose`
+
 - #310: Method for `OpenGraph`: `compose`
 
 - #277: Methods for pretty-printing `Pattern`: `to_ascii`,